+++ ga4gh/datamodel/__init__.py
@@ -1,0 +1,7 @@
+"""
+The GA4GH data model. Defines all the methods required to translate
+data in existing formats into GA4GH protocol types.
+"""
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
