+++ ga4gh/datamodel/reads.py
@@ -1,0 +1,14 @@
+"""
+Module responsible for translating read data into GA4GH native
+objects.
+"""
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
+
+
+import pysam
+
+import ga4gh.protocol as protocol
+
+# TODO use pysam to translate reads from SAM/BAM files into GA4GH objects.
