+++ ga4gh/datamodel/references.py
@@ -1,0 +1,13 @@
+"""
+Module responsible for translating reference sequence data into GA4GH native
+objects.
+"""
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
+
+
+import ga4gh.protocol as protocol
+
+# TODO use something to translate reference data in FASTA format into
+# GA4GH native objects.
